# docs/swagger.yaml
basePath: /api
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      id:
        type: integer
      modified_at:
        type: string
      modified_by:
        type: integer
      notes:
        type: string
      paid_at:
        type: string
      payment_method:
        type: string
      payment_type:
        description: down_payment, full_payment, deposit_in, deposit_refund, penalty_charge
        type: string
      reference_number:
        type: string
      rental_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.PaymentCreateDTO:
    properties:
      amount:
        type: number
      notes:
        type: string
      payment_method:
        type: string
      payment_type:
        type: string
      reference_number:
        type: string
      rental_id:
        type: integer
    required:
    - amount
    - payment_method
    - payment_type
    - rental_id
    type: object
  models.PaymentUpdateDTO:
    properties:
      amount:
        type: number
      notes:
        type: string
      payment_method:
        type: string
      payment_type:
        type: string
      reference_number:
        type: string
    type: object
...
info:
  contact:
    email: developer.haritsnb@gmail.com
    name: Harits Nala Barrun
  description: Backend RESTful API Sistem Kasir & Manajemen Rental Alat Camping menggunakan
    Golang + Gin + PostgreSQL + JWT Blacklist + KYC Upload.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Outdoor Gear Rental POS API
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Mencatat pembayaran tambahan/pelunasan manual untuk rental tertentu
      parameters:
      - description: Payload Pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PaymentCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Payment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Catat Pembayaran Manual
      tags:
      - Payments
  /payments/{id}:
    delete:
      description: Menonaktifkan data pembayaran (Admin Only)
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft Delete Pembayaran
      tags:
      - Payments
    get:
      description: Memuat detail satu bukti transaksi pembayaran
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Payment'
              type: object
        "401":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail Pembayaran Berdasarkan ID
      tags:
      - Payments
    post:
      description: Memulihkan kembali data pembayaran yang berstatus soft-deleted
        (Admin Only)
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore Pembayaran
      tags:
      - Payments
    put:
      consumes:
      - application/json
      description: Memperbarui nominal, metode, atau nomor referensi pembayaran (Admin
        Only)
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Update Pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PaymentUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Pembayaran (Flexible Partial)
      tags:
      - Payments
  /payments/{id}/force:
    delete:
      description: Menghapus fisik data pembayaran secara permanen dari database (Admin
        Only)
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Permanent Pembayaran (Force Delete)
      tags:
      - Payments
  /payments/deleted:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Payment'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Pembayaran Terhapus (Trash)
      tags:
      - Payments
  /payments/rental/{rental_id}:
    get:
      description: Memuat seluruh log pembayaran transaksi sewa (DP, pelunasan, deposit,
        refund, denda)
      parameters:
      - description: Rental ID
        in: path
        name: rental_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Payment'
                  type: array
              type: object
        "401":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Histori Pembayaran Berdasarkan Rental ID
      tags:
      - Payments
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"